+++ dapeng-code-generator/src/main/scala/com/github/dapeng/code/generator/JsonGenerator.scala
@@ -146,7 +146,7 @@ class JsonGenerator extends CodeGenerator {
     </div>
   }
 
-  override def toDataTypeTemplate(dataType:DataType): Elem = {
+  def toDataTypeTemplate(dataType:DataType): Elem = {
     dataType.kind match {
       case KIND.VOID => <div>void</div>
       case KIND.BOOLEAN => <div>boolean</div>
