+++ dapeng-code-generator/src/main/scala/com/github/dapeng/code/generator/ScalaGenerator.scala
@@ -737,7 +737,7 @@ class ScalaGenerator extends CodeGenerator {
     * @param dataType
     * @return
     */
-  override def toDataTypeTemplate(dataType:DataType): Elem = {
+  def toDataTypeTemplate(dataType:DataType): Elem = {
     dataType.kind match {
       case KIND.VOID => <div>Unit</div>
       case KIND.BOOLEAN => <div>Boolean</div>
